+++ src/translations/en.js
@@ -4,8 +4,10 @@ export const en = {
 	error: {
 		_: 'Error',
 		type: 'Incorrect value, expecting a __type__',
-		min: 'Value should be more than __minLength__ characters long',
-		max: 'Value should be less than __maxLength__ characters long',
+		min: 'Value must be more than __min__',
+		max: 'Value must be less than __max__',
+		minLength: 'Value must be more than __minLength__ characters long',
+		maxLength: 'Value must be less than __maxLength__ characters long',
 		required: 'Required',
 		minChildren: 'Expected to have at least __minChildren__',
 		maxChildren: 'Can only have __maxChildren__',