+++ src/translations/es.js
@@ -4,8 +4,10 @@ export const es = {
   error: {
     _: 'Error',
     type: 'Valor incorrecto, esperando __type__',
-    min: 'El valor debería ser superior a __min__',
-    max: 'El valor debería ser inferior a __max__',
+    min: 'El valor debe ser superior a __min__',
+    max: 'El valor debe ser inferior a __max__',
+		minLength: 'El valor debe tener como mínimo __minLength__ caracteres',
+		maxLength: 'El valor debe tener como máximo __maxLength__ caracteres',
     required: 'Requerido',
     minChildren: 'Se esperan __minChildren__',
     maxChildren: 'Sólo puede haber __maxChildren__',